feat: Add formatter utils for date and price
This commit adds a new file `formatter.ts` to the `src/utils` directory. The file contains two utility functions: `dateFormatter` and `priceFormatter`.

`dateFormatter` is an instance of `Intl.DateTimeFormat` configured with the locale `pt-BR`. It will be used to format dates in the Brazilian Portuguese format.

`priceFormatter` is an instance of `Intl.NumberFormat` configured with the locale `pt-BR` and the currency format for Brazilian Real (BRL). It will be used to format prices in the Brazilian currency format.

These formatter utilities will be used in the `Transactions` and `Summary` components to format dates and prices respectively.